use the new_ast as the default one
