+++ psydac/api/ast/__init__.py
@@ -1,6 +1,8 @@
-from psydac.api.ast import fem
-from psydac.api.ast import basic
-from psydac.api.ast import evaluation
+from psydac.api.ast import nodes
+from psydac.api.ast import parser
 from psydac.api.ast import utilities
+from psydac.api.ast import fem
 from psydac.api.ast import glt
 from psydac.api.ast import expr
+from psydac.api.ast import evaluation
+
