+++ psydac/api/ast/fem.py
@@ -1,491 +1,276 @@
-from collections import OrderedDict
-from itertools import groupby
-import numpy as np
+from .nodes import TensorQuadrature
+from .nodes import TensorBasis
+from .nodes import GlobalTensorQuadrature
+from .nodes import LocalTensorQuadrature
+from .nodes import LocalTensorQuadratureBasis
+from .nodes import LocalTensorQuadratureTestBasis
+from .nodes import LocalTensorQuadratureTrialBasis
+from .nodes import GlobalTensorQuadratureTestBasis
+from .nodes import GlobalTensorQuadratureTrialBasis
+from .nodes import TensorQuadratureBasis
+from .nodes import IndexElement, IndexQuadrature
+from .nodes import IndexDof, IndexDofTest, IndexDofTrial
+from .nodes import IndexDerivative
+from .nodes import LengthElement, LengthQuadrature
+from .nodes import LengthDof, LengthDofTest, LengthDofTrial
+from .nodes import SplitArray
+from .nodes import Reduction
+from .nodes import Reset
+from .nodes import LogicalValueNode
+from .nodes import TensorIteration
+from .nodes import TensorIterator
+from .nodes import TensorGenerator
+from .nodes import ProductIteration
+from .nodes import ProductIterator
+from .nodes import ProductGenerator
+from .nodes import StencilMatrixLocalBasis
+from .nodes import StencilMatrixGlobalBasis
+from .nodes import BlockStencilMatrixLocalBasis
+from .nodes import BlockStencilMatrixGlobalBasis
+from .nodes import BlockStencilVectorLocalBasis
+from .nodes import BlockStencilVectorGlobalBasis
+from .nodes import BlockMatrixNode
+from .nodes import StencilVectorLocalBasis
+from .nodes import StencilVectorGlobalBasis
+from .nodes import GlobalElementBasis
+from .nodes import LocalElementBasis
+from .nodes import TensorQuadratureTestBasis, TensorQuadratureTrialBasis
+from .nodes import TensorTestBasis,TensorTrialBasis
+from .nodes import GlobalSpan
+from .nodes import CoefficientBasis
+from .nodes import MatrixLocalBasis, MatrixGlobalBasis
+from .nodes import GeometryExpressions
+from .nodes import Loop
+from .nodes import EvalMapping, EvalField
+from .nodes import WeightedVolumeQuadrature
+from .nodes import ComputeLogical, ComputeKernelExpr
+from .nodes import ElementOf, Reduce
+from .nodes import construct_logical_expressions
+from .nodes import ComputePhysicalBasis
+from .nodes import Pads, Mask
 
-from sympy import symbols, Symbol, IndexedBase
-from sympy import Mul, Tuple
+from .utilities import physical2logical
+#==============================================================================
+from sympy import Basic
 from sympy import Matrix, ImmutableDenseMatrix
-from sympy import Mod, Abs
-from sympy.core.function import AppliedUndef
+from sympy import symbols
+from sympy.core.containers import Tuple
 
-from pyccel.ast.core import Variable, IndexedVariable
-from pyccel.ast.core import For
-from pyccel.ast.core import Assign
-from pyccel.ast.core import AugAssign
-from pyccel.ast.core import Slice
-from pyccel.ast.core import Range, Product
-from pyccel.ast.core import FunctionDef
-from pyccel.ast.core import FunctionCall
-from pyccel.ast import Zeros
-from pyccel.ast import Import
-from pyccel.ast import DottedName
-from pyccel.ast import Nil
-from pyccel.ast import Len
-from pyccel.ast import If, Is, Return
-from pyccel.ast import String, Print, Shape
-from pyccel.ast import Comment
-from pyccel.ast.core      import _atomic
-from pyccel.ast.utilities import build_types_decorator
+from sympde.expr import TerminalExpr
+from sympde.expr import LinearForm
+from sympde.expr import BilinearForm
+from sympde.expr import Functional
 
-from sympde.core                 import Constant
+from sympde.topology.basic       import Boundary
+from sympde.topology             import Mapping
+from sympde.topology             import H1SpaceType, HcurlSpaceType
+from sympde.topology             import HdivSpaceType, L2SpaceType, UndefinedSpaceType
+from sympde.topology             import element_of
 from sympde.topology             import ScalarField
 from sympde.topology             import VectorField, IndexedVectorField
-from sympde.topology             import Boundary, BoundaryVector, NormalVector, TangentVector
-from sympde.topology             import ElementArea
-from sympde.topology             import LogicalExpr
-from sympde.topology             import SymbolicExpr
-from sympde.topology             import UndefinedSpaceType
-from sympde.topology.space       import ScalarFunctionSpace, VectorFunctionSpace
-from sympde.topology.space       import ProductSpace
 from sympde.topology.space       import ScalarTestFunction
 from sympde.topology.space       import VectorTestFunction
-from sympde.topology.space       import element_of
 from sympde.topology.space       import IndexedTestTrial
 from sympde.topology.derivatives import _partial_derivatives
 from sympde.topology.derivatives import _logical_partial_derivatives
 from sympde.topology.derivatives import get_max_partial_derivatives
-from sympde.expr                 import BilinearForm, LinearForm, Functional
-
-from psydac.fem.splines import SplineSpace
-from psydac.fem.tensor  import TensorFemSpace
-from psydac.fem.vector  import ProductFemSpace
-
-from .basic      import SplBasic
-from .evaluation import EvalQuadratureMapping, EvalQuadratureField, EvalQuadratureVectorField
-from .utilities  import random_string
-from .utilities  import build_pythran_types_header, variables
-from .utilities  import compute_boundary_jacobian
-from .utilities  import compute_normal_vector, compute_tangent_vector
-from .utilities  import select_loops, filter_product
-from .utilities  import compute_atoms_expr
-from .utilities  import is_scalar_field, is_vector_field
-from .utilities  import math_atoms_as_str
 
 
-FunctionalForms = (BilinearForm, LinearForm, Functional)
+
+from pyccel.ast       import EmptyLine
+from pyccel.ast.core  import _atomic
+
+from .nodes import index_quad
+from .nodes import index_dof, index_dof_test, index_dof_trial
+from .nodes import index_element
+from .nodes import index_deriv
+
+from collections import OrderedDict
+from itertools   import groupby
+
+
 
 #==============================================================================
-def init_loop_quadrature(indices, ranges, boundary):
-    stmts = []
-    if not boundary:
-        return stmts
-
-    if isinstance(boundary, Boundary):
-        quad_mask = [boundary.axis]
-        quad_ext  = [boundary.ext]
-    else:
-        raise TypeError(boundary)
-
-    dim = len(indices)
-    for i in range(dim-1,-1,-1):
-        rx = ranges[i]
-        x = indices[i]
-
-        if i in quad_mask:
-            i_index = quad_mask.index(i)
-            ext = quad_ext[i_index]
-
-            stmts += [Assign(x, 0)]
-
-    return stmts
-
+def convert(dtype):
+    if isinstance(dtype, (H1SpaceType, UndefinedSpaceType)):
+        return 0
+    elif isinstance(dtype, HcurlSpaceType):
+        return 1
+    elif isinstance(dtype, HdivSpaceType):
+        return 2
+    elif isinstance(dtype, L2SpaceType):
+        return 3
 #==============================================================================
-def init_loop_basis(indices, ranges, boundary):
-    stmts = []
-    if not boundary:
-        return stmts
+def regroupe(tests):
+    tests  = [i.base if isinstance(i, IndexedTestTrial) else i for i in tests]
+    new_tests = []
+    for i in tests:
+        if i not in new_tests:
+            new_tests.append(i)
+    tests = new_tests
 
-    if isinstance(boundary, Boundary):
-        quad_mask = [boundary.axis]
-        quad_ext  = [boundary.ext]
-    else:
-        raise TypeError(boundary)
-
-    dim = len(indices)
-    for i in range(dim-1,-1,-1):
-        rx = ranges[i]
-        x = indices[i]
-
-        if i in quad_mask:
-            i_index = quad_mask.index(i)
-            ext = quad_ext[i_index]
-
-            if ext == -1:
-                value = rx.start
-
-            elif ext == 1:
-                value = rx.stop - 1
-
-            stmts += [Assign(x, value)]
-
-    return stmts
-
-#==============================================================================
-def init_loop_support(indices_elm, n_elements,
-                      indices_span, spans, ranges,
-                      points_in_elm, points,
-                      weights_in_elm, weights,
-                      test_basis_in_elm, test_basis,
-                      trial_basis_in_elm, trial_basis,
-                      is_bilinear, boundary):
-    stmts = []
-    if not boundary:
-        return stmts
-
-    if isinstance(boundary, Boundary):
-        quad_mask = [boundary.axis]
-        quad_ext  = [boundary.ext]
-    else:
-        raise TypeError(boundary)
-
-    dim = len(indices_elm)
-    for i in range(dim-1,-1,-1):
-        rx = ranges[i]
-        x = indices_elm[i]
-
-        if i in quad_mask:
-            i_index = quad_mask.index(i)
-            ext = quad_ext[i_index]
-
-            if ext == -1:
-                value = rx.start
-
-            elif ext == 1:
-                value = rx.stop - 1
-
-            stmts += [Assign(x, value)]
-
-    axis = quad_mask[0]
-
-    # ... assign element index
-    ncells = n_elements[axis]
-    ie = indices_elm[axis]
-    # ...
-
-    # ... assign span index
-    i_span = indices_span[axis]
-    stmts += [Assign(i_span, spans[axis][ie])]
-    # ...
-
-    # ... assign points, weights and basis
-    # ie is substitute by 0
-    # sympy does not like ':'
-    _slice = Slice(None,None)
-
-    stmts += [Assign(points_in_elm[axis], points[axis][0,_slice])]
-    stmts += [Assign(weights_in_elm[axis], weights[axis][0,_slice])]
-    stmts += [Assign(test_basis_in_elm[axis], test_basis[axis][0,_slice,_slice,_slice])]
-
-    if is_bilinear:
-        stmts += [Assign(trial_basis_in_elm[axis], trial_basis[axis][0,_slice,_slice,_slice])]
-    # ...
-
-    return stmts
-
-#==============================================================================
-# TODO take exponent to 1/dim
-def area_eval_mapping(mapping, area, dim, indices_quad, weight):
-
-    stmts = []
-
-    # mapping components and their derivatives
-    ops      = _logical_partial_derivatives[:dim]
-    elements = [d(mapping[i]) for d in ops for i in range(0, dim)]
-
-    # declarations
-    stmts += [Comment('declarations')]
-    for e in elements:
-        lhs      = SymbolicExpr(e)
-        rhs_name = lhs.name + '_values'
-        rhs      = IndexedBase(rhs_name)[indices_quad]
-        stmts   += [Assign(lhs, rhs)]
-
-    # jacobian determinant
-    jac    = SymbolicExpr(mapping.det_jacobian)
-    stmts += [AugAssign(area, '+', Abs(jac) * weight)]
-
-    return stmts
-
-#==============================================================================
-# target is used when there are multiple expression (domain/boundaries)
-class Kernel(SplBasic):
-    """
-    Generate the AST of a function for computing an integral form over a
-    single domain element, or boundary element.
-
-    For a bilinear form, such a function will compute an 'element matrix'.
-    For a linear form, it will compute an 'element vector'.
-    For a functional, it will compute an 'element value' (scalar).
-
-    """
-    def __new__(cls, weak_form, kernel_expr, target=None,
-                boundary=None, name=None, boundary_basis=None,
-                mapping=None, is_rational_mapping=None,symbolic_space=None, backend=None):
-
-        if not isinstance(weak_form, FunctionalForms):
-            raise TypeError('> Expecting a weak formulation')
-
-        if symbolic_space:
-            symbolic_space= symbolic_space[0]
-            
-        unique_scalar_space = True
-        if isinstance(symbolic_space, ProductSpace):
-            spaces = symbolic_space.spaces
-            space = spaces[0]
-            unique_scalar_space = all(sp.kind==space.kind for sp in spaces)
-        elif isinstance(symbolic_space, VectorFunctionSpace):
-            unique_scalar_space = isinstance(symbolic_space.kind, UndefinedSpaceType)
-
-        # ...
-        # get the target expr if there are multiple expressions (domain/boundary)
-        on_boundary = False
-        if target is None:
-            if len(kernel_expr) > 1:
-                msg = '> weak form has multiple expression, but no target was given'
-                raise ValueError(msg)
-
-            e = kernel_expr[0]
-            on_boundary = isinstance(e.target, Boundary)
-            kernel_expr = e.expr
-
+    spaces = [i.space for i in tests]
+    kinds  = [i.kind for i in spaces]
+    funcs  = OrderedDict(zip(tests, kinds))
+    funcs  = sorted(funcs.items(), key=lambda x:convert(x[1]))
+    grs = [OrderedDict(g) for k,g in groupby(funcs,key=lambda x:convert(x[1]))]
+    grs = [(list(g.values())[0],tuple(g.keys())) for g in grs]
+    groups = []
+    for d,g in grs:
+        if isinstance(d, (HcurlSpaceType, HdivSpaceType)) and isinstance(g[0], VectorTestFunction):
+            dim = g[0].space.ldim
+            for i in range(dim):
+                s = [u[i] for u in g]
+                groups += [(d,tuple(s))]
         else:
-            ls = [i for i in kernel_expr if i.target is target]
-            e = ls[0]
-            on_boundary = isinstance(e.target, Boundary)
-            kernel_expr = e.expr
-        # ...
-
-        # ...
-        if boundary:
-            if not isinstance(boundary, Boundary):
-                raise TypeError('> Expecting a Boundary for boundary')
-        # ...
-
-        # ... boundary must be given if there are Trace nodes
-        if on_boundary and not boundary:
-            raise ValueError('> boundary must be provided for a boundary Kernel')
-        # ...
-
-        # ... default value for boundary_basis is True if on boundary
-        if on_boundary and (boundary_basis is None):
-            boundary_basis = True
-        # ...
-
-        tag = random_string( 8 )
-        obj = SplBasic.__new__(cls, tag, name=name,
-                               prefix='kernel', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._weak_form           = weak_form
-        obj._kernel_expr         = kernel_expr
-        obj._target              = target
-        obj._boundary            = boundary
-        obj._boundary_basis      = boundary_basis
-        obj._area                = None
-        obj._user_functions      = []
-        obj._backend             = backend
-        obj._symbolic_space      = symbolic_space
-        obj._unique_scalar_space = unique_scalar_space
-
-        obj._func = obj._initialize()
-        return obj
-
-    @property
-    def weak_form(self):
-        return self._weak_form
-
-    @property
-    def kernel_expr(self):
-        return self._kernel_expr
-
-    @property
-    def target(self):
-        return self._target
-
-    @property
-    def boundary_basis(self):
-        return self._boundary_basis
-
-    @property
-    def n_rows(self):
-        return self._n_rows
-
-    @property
-    def n_cols(self):
-        return self._n_cols
-
-    @property
-    def max_nderiv(self):
-        return self._max_nderiv
-
-    @property
-    def zero_terms(self):
-        return self._zero_terms
-
-    @property
-    def constants(self):
-        return self._constants
-
-    @property
-    def fields(self):
-        return self._fields
-
-    @property
-    def fields_coeffs(self):
-        return self._fields_coeffs
-
-    @property
-    def vector_fields(self):
-        return self._vector_fields
-
-    @property
-    def vector_fields_coeffs(self):
-        return self._vector_fields_coeffs
-
-    @property
-    def fields_val(self):
-        return self._fields_val
-
-    @property
-    def vector_fields_val(self):
-        return self._vector_fields_val
-
-    @property
-    def mapping_coeffs(self):
-        if not self.eval_mapping:
-            return ()
-
-        return self.eval_mapping.mapping_coeffs
-
-    @property
-    def mapping_values(self):
-        if not self.eval_mapping:
-            return ()
-
-        return self.eval_mapping.mapping_values
-
-    @property
-    def eval_fields(self):
-        return self._eval_fields
-
-    @property
-    def eval_vector_fields(self):
-        return self._eval_vector_fields
-
-    @property
-    def eval_mapping(self):
-        return self._eval_mapping
-
-    @property
-    def area(self):
-        return self._area
-
-    @property
-    def user_functions(self):
-        return self._user_functions
-
-    @property
-    def unique_scalar_space(self):
-        return self._unique_scalar_space
-        
-    @property
-    def symbolic_space(self):
-        return self._symbolic_space
-        
-    @property
-    def backend(self):
-        return self._backend
-
-    def build_arguments(self, data):
-
-        other = data
-
-        if self.mapping_values:
-            other = self.mapping_values + other
-
-        if self.constants:
-            other = other + self.constants
-
-        return self.basic_args + other
-
-    def _initialize(self):
-        is_linear   = isinstance(self.weak_form, LinearForm)
-        is_bilinear = isinstance(self.weak_form, BilinearForm)
-        is_function = isinstance(self.weak_form, Functional)
-        unique_scalar_space = self.unique_scalar_space
-
-        expr = self.kernel_expr
-        mapping = self.mapping
-
-        # ... area of an element
-        area = list(expr.atoms(ElementArea))
-        if area:
-            assert(len(area) == 1)
-            area = area[0]
-
-            self._area = Variable('real', 'area')
-
-            # update exp
-            expr = expr.subs(area, self.area)
-        # ...
-
-        # ... undefined functions
-        funcs = expr.atoms(AppliedUndef)
-        if funcs:
-            self._user_functions = [f.func for f in list(funcs)]
-        # ...
-
-        # ...
-        n_rows = 1 ; n_cols = 1
-        if is_bilinear:
-            if isinstance(expr, (Matrix, ImmutableDenseMatrix)):
-                n_rows = expr.shape[0]
-                n_cols = expr.shape[1]
-
-        if is_linear:
-            if isinstance(expr, (Matrix, ImmutableDenseMatrix)):
-                n_rows = expr.shape[0]
-
-        self._n_rows = n_rows
-        self._n_cols = n_cols
-        # ...
-
-        dim      = self.weak_form.ldim
-        dim_test = dim
-
-        if is_bilinear:
-            dim_trial = dim
+            groups += [(d,g)]
+    return groups
+#==============================================================================
+def expand(args):
+    new_args = []
+    for i in args:
+        if isinstance(i, (ScalarTestFunction, IndexedTestTrial)):
+            new_args += [i]
+        elif isinstance(i, VectorTestFunction):
+            new_args += [i[k] for k in  range(i.space.ldim)]
         else:
-            dim_trial = 0
+            raise NotImplementedError("TODO")
+    return new_args
+#==============================================================================
+def expand_hdiv_hcurl(args):
+    new_args = []
+    for i in args:
+        if isinstance(i, (ScalarTestFunction, IndexedTestTrial)):
+            new_args += [i]
+        elif isinstance(i, VectorTestFunction):
+            if isinstance(i.space.kind, (HcurlSpaceType, HdivSpaceType)):
+                new_args += [i[k] for k in  range(i.space.ldim)]
+            else:
+                new_args += [i]
+        else:
+            raise NotImplementedError("TODO")
+    return new_args
+#==============================================================================
+def get_length(args):
+    n = 0
+    for i in args:
+        if isinstance(i, ScalarTestFunction):
+            n += 1
+        elif isinstance(i, VectorTestFunction):
+            n += i.space.ldim
+        else:
+            raise NotImplementedError("TODO")
+    return n
 
-        # ... coordinates
-        coordinates = self.weak_form.coordinates
-        if dim == 1:
-            coordinates = [coordinates]
+#==============================================================================
+def is_scalar_field(expr):
+
+    if isinstance(expr, _partial_derivatives):
+        return is_scalar_field(expr.args[0])
+
+    elif isinstance(expr, _logical_partial_derivatives):
+        return is_scalar_field(expr.args[0])
+
+    elif isinstance(expr, ScalarField):
+        return True
+
+    return False
+
+#==============================================================================
+def is_vector_field(expr):
+
+    if isinstance(expr, _partial_derivatives):
+        return is_vector_field(expr.args[0])
+
+    elif isinstance(expr, _logical_partial_derivatives):
+        return is_vector_field(expr.args[0])
+
+    elif isinstance(expr, (VectorField, IndexedVectorField)):
+        return True
+
+    return False
+#==============================================================================
+class Block(Basic):
+    """."""
+    def __new__(cls, body):
+        if not isinstance(body, (list, tuple, Tuple)):
+            body = [body]
+
+        body = Tuple(*body)
+        return Basic.__new__(cls, body)
+
+    @property
+    def body(self):
+        return self._args[0]
+#==============================================================================
+class DefNode(Basic):
+    """."""
+    def __new__(cls, name, arguments, local_variables, body):
+        return Basic.__new__(cls, name, arguments, local_variables, body)
+
+    @property
+    def name(self):
+        return self._args[0]
+
+    @property
+    def arguments(self):
+        return self._args[1]
+
+    @property
+    def local_variables(self):
+        return self._args[2]
+
+    @property
+    def body(self):
+        return self._args[3]
+#==============================================================================
+class AST(object):
+    """
+    """
+    def __init__(self, expr, terminal_expr, spaces, is_rational_mapping=False, tag=None):
+        # ... compute terminal expr
+        # TODO check that we have one single domain/interface/boundary
+
+        is_bilinear   = False
+        is_linear     = False
+        is_functional = False
+        tests         = []
+        trials        = []
         # ...
+        #TODO use LogicalExpr instead and return the appropriate expression in case of analytical or non analytical mapping
 
-        # ...
-        constants = tuple(expr.atoms(Constant))
-        self._constants = []
-        # we need this, since Constant is an extension of Symbol and the type is
-        # given as for sympy Symbol
-        for c in constants:
-            dtype = 'real'
-            if c.is_integer:
-                dtype = 'int'
+        domain        = terminal_expr.target
+        terminal_expr = terminal_expr.expr
+        dim           = domain.dim
+        mask          = None
+        if isinstance(domain, Boundary):
+            mask = Mask(domain.axis, domain.ext)
 
-            elif c.is_real:
-                dtype = 'real'
+        if isinstance(expr, LinearForm):
+            is_linear = True
+            tests     = expr.test_functions
+            mapping   = spaces.symbolic_mapping
+            spaces    = spaces.symbolic_space
 
-            elif c.is_complex:
-                dtype = 'complex'
 
-            self._constants.append(Variable(dtype, str(c.name)))
+        elif isinstance(expr, BilinearForm):
+            is_bilinear = True
+            tests       = expr.test_functions
+            trials      = expr.trial_functions
+            mapping     = spaces[0].symbolic_mapping
+            spaces      = [V.symbolic_space for V in spaces]
 
-        self._constants = tuple(self._constants)
-        # ...
+        elif isinstance(expr, Functional):
+            is_functional = True
+            fields = tuple(expr.atoms(ScalarTestFunction, VectorTestFunction))
+            assert len(fields) == 1
+            tests   = fields[0]
+            tests   = Tuple(tests)
+            mapping = spaces.symbolic_mapping
+            spaces  = spaces.symbolic_space
+        else:
+            raise NotImplementedError('TODO')
 
-        # ...
         atoms_types = (_partial_derivatives,
                        VectorTestFunction,
                        ScalarTestFunction,
@@ -493,1651 +278,453 @@ class Kernel(SplBasic):
                        ScalarField,
                        VectorField, IndexedVectorField)
 
-        atoms  = _atomic(expr, cls=atoms_types)
-        # ...
-
-        # ...
-        atomic_expr_field        = [atom for atom in atoms if is_scalar_field(atom)]
-        atomic_expr_vector_field = [atom for atom in atoms if is_vector_field(atom)]
-
-        atomic_expr       = [atom for atom in atoms if not( atom in atomic_expr_field ) and
-                                                       not( atom in atomic_expr_vector_field)]
-        # ...
-
-
-        field_atoms   = tuple(expr.atoms(ScalarField))
-        fields        = []
-
-        # ... create EvalQuadratureField
-        self._eval_fields = []
-        self._map_stmts_fields = OrderedDict()
-        if atomic_expr_field:
-            keyfunc = lambda F: F.space.name
-            data = sorted(field_atoms, key=keyfunc)
-            for space_str, group in groupby(data, keyfunc):
-                g_names = set([f.name for f in group])
-                fields_expressions = []
-                for e in atomic_expr_field:
-                    fs = e.atoms(ScalarField)
-                    f_names = set([f.name for f in fs])
-                    if f_names & g_names:
-                        fields_expressions += [e]
-                        space = list(fs)[0].space
-
-                eval_field = EvalQuadratureField(space, fields_expressions,
-                                       boundary       = self.boundary,
-                                       boundary_basis = self.boundary_basis,
-                                       mapping        = mapping,
-                                       backend        = self.backend)
-
-                fields += list(eval_field.fields)
-                self._eval_fields.append(eval_field)
-                for k,v in eval_field.map_stmts.items():
-                    self._map_stmts_fields[k] = v
-
-        # update dependencies
-        self._dependencies += self.eval_fields
-
-        # TODO: remove these?
-        d_subs = dict(zip(_partial_derivatives, _logical_partial_derivatives))
-        fields_logical     = tuple(f.subs(d_subs) for f in fields)
-        fields_str         = tuple(SymbolicExpr(f).name for f in fields)
-        fields_logical_str = tuple(SymbolicExpr(f).name for f in fields_logical)
-        # ...
-
-        vector_field_atoms   = tuple(expr.atoms(VectorField))
-        vector_fields        = []
-
-        # ... create EvalQuadratureVectorField
-        self._eval_vector_fields = []
-        if atomic_expr_vector_field:
-            keyfunc = lambda F: F.space.name
-            data = sorted(vector_field_atoms, key=keyfunc)
-            for space_str, group in groupby(data, keyfunc):
-                g_names = set([f.name for f in group])
-                vector_fields_expressions = []
-                for e in atomic_expr_vector_field:
-                    fs = e.atoms(VectorField)
-                    f_names = set([f.name for f in fs])
-                    if f_names & g_names:
-                        vector_fields_expressions += [e]
-                        space = list(fs)[0].space
-
-                eval_vector_field = EvalQuadratureVectorField(space, vector_fields_expressions,
-                                                    boundary       = self.boundary,
-                                                    boundary_basis = self.boundary_basis,
-                                                    mapping        = mapping,
-                                                    backend        = self.backend)
-
-                vector_fields += list(eval_vector_field.vector_fields)
-                self._eval_vector_fields.append(eval_vector_field)
-                for k,v in eval_vector_field.map_stmts.items():
-                    self._map_stmts_fields[k] = v
-
-        # update dependencies
-        self._dependencies  += self.eval_vector_fields
-
-        # TODO: remove these?
-        vector_fields_logical     = tuple(f.subs(d_subs) for f in vector_fields)
-        vector_fields_str         = tuple(SymbolicExpr(f).name for f in vector_fields)
-        vector_fields_logical_str = tuple(SymbolicExpr(f).name for f in vector_fields_logical)
-        # ...
-
-        # ... TODO add it as a method to basic class
         nderiv = 1
-        if isinstance(self.kernel_expr, Matrix):
-            n_rows, n_cols = self.kernel_expr.shape
+        if isinstance(terminal_expr, Matrix):
+            n_rows, n_cols = terminal_expr.shape
+            atomic_expr = terminal_expr.zeros(n_rows, n_cols)
+            atomic_expr_field = []
             for i_row in range(0, n_rows):
                 for i_col in range(0, n_cols):
-                    d = get_max_partial_derivatives(self.kernel_expr[i_row,i_col])
+                    d = get_max_partial_derivatives(terminal_expr[i_row,i_col])
                     nderiv = max(nderiv, max(d.values()))
+                    atoms  = _atomic(terminal_expr[i_row, i_col], cls=atoms_types)
+                    fields = [atom for atom in atoms if is_scalar_field(atom) or is_vector_field(atom)]
+                    atoms  = [atom for atom in atoms if atom not in fields ]
+                    atomic_expr[i_row, i_col] = Tuple(*atoms)
+                    atomic_expr_field += [*fields]
         else:
-            d = get_max_partial_derivatives(self.kernel_expr)
+            d = get_max_partial_derivatives(terminal_expr)
             nderiv = max(nderiv, max(d.values()))
+            atoms  = _atomic(terminal_expr, cls=atoms_types)
+            atomic_expr_field = [atom for atom in atoms if is_scalar_field(atom) or is_vector_field(atom)]
+            atomic_expr       = [atom for atom in atoms if atom not in atomic_expr_field ]
+            atomic_expr       = Matrix([[Tuple(*atomic_expr)]])
+            terminal_expr     = Matrix([[terminal_expr]])
 
-        self._max_nderiv = nderiv
-        # ...
+        # ...
 
-        # ... mapping
-        mapping = self.mapping
-        self._eval_mapping = None
-        if mapping:
+        d_tests = {}
+        tests  = expand_hdiv_hcurl(tests)
+        trials = expand_hdiv_hcurl(trials)
+        for v in tests:
+            d = {}
+            d['global'] = GlobalTensorQuadratureTestBasis(v)
+            d['local']  = LocalTensorQuadratureTestBasis(v)
+            d['array']  = TensorQuadratureTestBasis(v)
+            d['basis']  = TensorTestBasis(v)
+            d['span']   = GlobalSpan(v)
 
-            if is_bilinear or is_linear:
-                space = self.weak_form.test_spaces[0]
+            d_tests[v] = d
+        # ...
 
-            elif is_function:
-                space = self.weak_form.space
+        # ...
+        d_trials = {}
+        for v in trials:
+            d = {}
+            d['global'] = GlobalTensorQuadratureTrialBasis(v)
+            d['local']  = LocalTensorQuadratureTrialBasis(v)
+            d['array']  = TensorQuadratureTrialBasis(v)
+            d['basis']  = TensorTrialBasis(v)
+            d['span']   = GlobalSpan(v)
 
-            eval_mapping = EvalQuadratureMapping(space, mapping,
-                                       boundary       = self.boundary,
-                                       boundary_basis = self.boundary_basis,
-                                       nderiv         = nderiv,
-                                       is_rational_mapping = self.is_rational_mapping,
-                                       area           = self.area,
-                                       backend        = self.backend)
+            d_trials[v] = d
+        # ...
 
-            self._eval_mapping = eval_mapping
+        shapes_tests  = {}
+        shapes_trials = {}
 
-            # update dependencies
-            self._dependencies += [self.eval_mapping]
-        # ...
+        start = 0
+        for v in tests:
+            ln = 1 if isinstance(v, ScalarTestFunction) else dim
+            end = start + ln
+            shapes_tests[v] = (start, end)
+            start = end
 
-        if is_bilinear or is_linear:
-            test_function = self.weak_form.test_functions
-            if not isinstance(test_function, (tuple, Tuple)):
-                test_function = Tuple(test_function)
+        start = 0
+        for u in trials:
+            ln = 1 if isinstance(u, ScalarTestFunction) else dim
+            end = start + ln
+            shapes_trials[u] = (start, end)
+            start = end
 
-        elif is_function:
-            test_function = element_of(self.weak_form.space, name='Nj')
-            test_function = Tuple(test_function)
+        # ...
+        if is_linear:
+            ast = _create_ast_linear_form(terminal_expr, atomic_expr, atomic_expr_field, 
+                                          tests, d_tests, shapes_tests,
+                                          nderiv, domain.dim, 
+                                          mapping, spaces, mask, tag)
 
-        # creation of symbolic vars
-        if is_bilinear:
-            rank = 2*dim
+        elif is_bilinear:
+            ast = _create_ast_bilinear_form(terminal_expr, atomic_expr, atomic_expr_field,
+                                            tests, d_tests, shapes_tests,
+                                            trials, d_trials, shapes_trials,
+                                            nderiv, domain.dim, 
+                                            mapping, spaces, mask, tag)
 
-        elif is_linear:
-            rank = dim
-
-        elif is_function:
-            rank = 1
-
-        if isinstance(expr, Matrix):
-            nr, nc = expr.shape
-            mats = ['mat_{}{}'.format(i, j) for j in range(nc) for i in range(nr)]
-            vals = ['val_{}{}'.format(i, j) for j in range(nc) for i in range(nr)]
-
-            mats = variables(mats, dtype='real', rank=rank, cls=IndexedVariable)
-            vals = variables(vals, 'real')
-            expr = expr[:]
-            ln   = len(expr)
-
+        elif is_functional:
+            ast = _create_ast_functional_form(terminal_expr, atomic_expr_field,
+                                              tests, d_tests, shapes_tests,
+                                              nderiv, domain.dim, 
+                                              mapping, spaces, mask, tag)
         else:
-            mats = (IndexedVariable('mat_00', dtype='real', rank=rank),)
-            vals = (Variable('real', 'val_00'),)
-            expr = [expr]
-            ln   = 1
-            
-        # ... looking for 0 terms
-        zero_terms = [i for i,e in enumerate(expr) if e == 0]
-        self._zero_terms = zero_terms
-        
+            raise NotImplementedError('TODO')
         # ...
 
-        # ... declarations
-        fields        = symbols(fields_str)
-        fields_logical = symbols(fields_logical_str)
+        self._expr    = ast
+        self._nderiv  = nderiv
+        self._domain  = domain
+        self._mapping = mapping
 
-        fields_coeffs = variables(['coeff_{}'.format(f) for f in field_atoms],
-                                          dtype='real', rank=dim, cls=IndexedVariable)
-        fields_val    = variables(['{}_values'.format(f) for f in fields_logical_str],
-                                          dtype='real', rank=dim, cls=IndexedVariable)
+    @property
+    def expr(self):
+        return self._expr
 
-        vector_fields        = symbols(vector_fields_str)
-        vector_fields_logical = symbols(vector_fields_logical_str)
+    @property
+    def nderiv(self):
+        return self._nderiv
 
-        vector_field_atoms = [f[i] for f in vector_field_atoms for i in range(0, dim)]
-        coeffs = ['coeff_{}'.format(SymbolicExpr(f).name) for f in vector_field_atoms]
-        vector_fields_coeffs = variables(coeffs, dtype='real', rank=dim, cls=IndexedVariable)
+    @property
+    def domain(self):
+        return self._domain
 
-        vector_fields_val    = variables(['{}_values'.format(f) for f in vector_fields_str],
-                                          dtype='real', rank=dim, cls=IndexedVariable)
+    @property
+    def mapping(self):
+        return self._mapping
 
-        test_degrees  = variables('test_d1:%s'%(dim+1),  'int')
-        trial_degrees = variables('trial_d1:%s'%(dim+1), 'int')
-        test_pads     = variables('test_p1:%s'%(dim+1),  'int')
-        trial_pads    = variables('trial_p1:%s'%(dim+1), 'int')
+    @property
+    def dim(self):
+        return self.domain.dim
+#==============================================================================
+#TODO add the parallel case
 
-        indices_quad  = variables('g1:%s'%(dim+1),  'int')
-        qds_dim       = variables('k1:%s'%(dim+1),  'int')
-        indices_test  = variables('il1:%s'%(dim+1), 'int')
-        indices_trial = variables('jl1:%s'%(dim+1), 'int')
-        wvol          = Variable('real', 'wvol')
+#================================================================================================================================
+def _create_ast_bilinear_form(terminal_expr, atomic_expr, atomic_expr_field,
+                              tests, d_tests, shapes_tests,
+                              trials, d_trials, shapes_trials,
+                              nderiv, dim, mapping, spaces, mask, tag):
+    """
+    """
 
-        basis_trial   = variables('trial_bs1:%s'%(dim+1),
-                                  dtype='real', rank=3, cls=IndexedVariable)
-        basis_test    = variables('test_bs1:%s'%(dim+1),
-                                  dtype='real', rank=3, cls=IndexedVariable)
-        weighted_vols = variables('quad_w1:%s'%(dim+1),
-                                  dtype='real', rank=1, cls=IndexedVariable)
-        positions     = variables('quad_u1:%s'%(dim+1),
-                                  dtype='real', rank=1, cls=IndexedVariable)
+    pads   = symbols('pad1, pad2, pad3')[:dim]
+    g_quad = GlobalTensorQuadrature()
+    l_quad = LocalTensorQuadrature()
+    #TODO should we use tests or trials for fields
+    coeffs   = [CoefficientBasis(i) for i in tests]
+    l_coeffs = [MatrixLocalBasis(i) for i in tests]
+    geo      = GeometryExpressions(mapping, nderiv)
 
-        # Used only if there is a mapping
-        inv_jac = Symbol('inv_jac')
-        det_jac = Symbol('det_jac')
+    l_mats  = BlockStencilMatrixLocalBasis(trials, tests, terminal_expr, dim, tag)
+    g_mats  = BlockStencilMatrixGlobalBasis(trials, tests, pads, terminal_expr, l_mats.tag)
 
-        # Used only in the case of a boundary assembly
-        vectors     = self.kernel_expr.atoms(BoundaryVector)
-        normal_vec  = symbols('normal_1:%d'%(dim+1))
-        tangent_vec = symbols('tangent_1:%d'%(dim+1))
-        det_jac_bnd = symbols('det_jac_bnd')
+    q_basis_tests  = OrderedDict((v,d_tests[v]['array'])  for v in tests)
+    q_basis_trials = OrderedDict((u,d_trials[u]['array']) for u in trials)
 
-        # ...
+    l_basis_tests  = OrderedDict((v,d_tests[v]['local'])  for v in tests)
+    l_basis_trials = OrderedDict((u,d_trials[u]['local']) for u in trials)
 
-        # ...
-        if is_bilinear:
-            self._basic_args = (test_degrees + trial_degrees + trial_pads +
-                                basis_test + basis_trial +
-                                positions + weighted_vols)
-                                
-            if self.eval_fields:
-                self._basic_args = self._basic_args + fields_val
+    g_basis_tests  = OrderedDict((v,d_tests[v]['global'])  for v in tests)
+    g_basis_trials = OrderedDict((u,d_trials[u]['global']) for u in trials)
 
-        if is_linear or is_function:
-            self._basic_args = (test_degrees +
-                                basis_test +
-                                positions + weighted_vols+
-                                fields_val + vector_fields_val)
-        # ...
+    g_span          = OrderedDict((u,d_tests[u]['span']) for u in tests)
 
-        # ...
-        if mapping:
-            mapping_elements = [SymbolicExpr(i) for i in self.eval_mapping.elements]
-            mapping_coeffs   = self.eval_mapping.mapping_coeffs
-            mapping_values   = self.eval_mapping.mapping_values
-        else:
-            mapping_elements = ()
-            mapping_coeffs   = ()
-            mapping_values   = ()
-        # ...
+    lengths_trials  = OrderedDict((u,LengthDofTrial(u)) for u in trials)
+    lengths_tests   = OrderedDict((v,LengthDofTest(v)) for v in tests)
+    # ...........................................................................................
+    quad_length = LengthQuadrature()
+    el_length   = LengthElement()
+    lengths     = [el_length, quad_length]
 
-        # ...
-        self._fields_val = fields_val
-        self._vector_fields_val = vector_fields_val
-        self._fields = fields
-        self._fields_logical = fields_logical
-        self._fields_coeffs = fields_coeffs
-        self._vector_fields = vector_fields
-        self._vector_fields_logical = vector_fields_logical
-        self._vector_fields_coeffs = vector_fields_coeffs
-        self._mapping_coeffs = mapping_coeffs
-        self._imports = set()
-        # ...
+    ind_quad      = index_quad.set_length(quad_length)
+    ind_element   = index_element.set_length(el_length)
+    ind_dof_test = index_dof_test.set_length(LengthDofTest(tests[0]))
+    # ...........................................................................................
+    eval_mapping = EvalMapping(ind_quad, ind_dof_test, q_basis_tests[tests[0]], l_basis_tests[tests[0]], mapping, geo, spaces[1], tests, nderiv, mask)
 
-        # ranges
-        ranges_test  = [Range(test_degrees[i]+1) for i in range(dim_test)]
-        ranges_trial = [Range(trial_degrees[i]+1) for i in range(dim_trial)]
-        ranges_quad  = [Range(qds_dim[i]) for i in range(dim)]
-        # ...
+    if atomic_expr_field:
+        raise NotImplementedError('TODO')
+        eval_field   = EvalField(atomic_expr_field, ind_quad, q_basis_trials[trials[0]], coeff, trials, mapping, nderiv)
 
-        # body of kernel
-        init_stmts, map_stmts = compute_atoms_expr(atomic_expr,
-                                                 indices_quad,
-                                                 indices_test,
-                                                 indices_trial,
-                                                 basis_trial,
-                                                 basis_test,
-                                                 coordinates,
-                                                 test_function,
-                                                 is_linear,
-                                                 mapping)
+    # ... loop over tests to evaluate fields
+    fields = EmptyLine()
+    if atomic_expr_field:
+        fields = Loop((*l_basis_trials, l_coeff), index_dof, [eval_field])
 
-        # Sort statements according to left-hand-side name
-        init_basis = sorted(init_stmts, key=lambda s: str(s.lhs))
-        init_map   = sorted( map_stmts, key=lambda s: str(s.lhs))
+    g_stmts = [eval_mapping, fields]
+    test_groups  = regroupe(tests)
+    trial_groups = regroupe(trials)
+    ex_tests     = expand(tests)
+    ex_trials    = expand(trials)
 
-        # Prepare matrix of functions for assembling each element of a block
-        # matrix or block vector.
-        if unique_scalar_space:
-            ln    = 1
-            funcs = [[None]]
-        else:
-            funcs = [[None] * self._n_cols] * self._n_rows
-
-        # Compute jacobian determinant expression
-        if mapping:
-            jac = SymbolicExpr(mapping.det_jacobian)
-
-        #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-        # FOR LOOP OVER THE EQUATIONS IN A SYSTEM
-        #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-        for indx in range(ln):
-
-            if not unique_scalar_space and indx in zero_terms:
+    #=========================================================begin kernel======================================================
+    for te_dtype,sub_tests in test_groups:
+        for tr_dtype, sub_trials in trial_groups:
+            tests_indices = [ex_tests.index(i) for i in expand(sub_tests)]
+            trials_indices = [ex_trials.index(i) for i in expand(sub_trials)]
+            sub_terminal_expr = terminal_expr[tests_indices,trials_indices]
+            sub_atomic_expr   = atomic_expr[tests_indices,trials_indices]
+            l_sub_mats  = BlockStencilMatrixLocalBasis(sub_trials, sub_tests, sub_terminal_expr, dim, l_mats.tag)
+            if sub_terminal_expr.is_zero:
                 continue
+            q_basis_tests  = OrderedDict((v,d_tests[v]['array'])  for v in sub_tests)
+            q_basis_trials = OrderedDict((u,d_trials[u]['array']) for u in sub_trials)
 
-            elif not unique_scalar_space:
-                start = indx
-                end   = indx + 1
-                i_row = indx // self._n_cols
-                i_col = indx - i_row * self._n_cols
+            l_basis_tests  = OrderedDict((v,d_tests[v]['local'])  for v in sub_tests)
+            l_basis_trials = OrderedDict((u,d_trials[u]['local']) for u in sub_trials)
 
-            else:
-                i_row = 0
-                i_col = 0
-                start = 0
-                end   = len(expr)
-
-            #------------------------------------------------------------------
-            # Body of inner quadrature loop: all quantities are scalars
-            #------------------------------------------------------------------
-
-            # New body
-            body = []
-
-            # Evaluation of basis functions (and their derivatives)
-            # along each direction in the parametric domain
-            body += init_basis
-
-            # Evaluation of physical coordinates (x, y, z).
-            # In the case of a mapping, also compute the components
-            # of the Jacobian matrix, as well as its determinant.
-            if mapping:
-                body += [Assign(lhs, rhs[indices_quad])
-                         for lhs, rhs in zip(mapping_elements, mapping_values)]
-
-                body += [Assign(det_jac, jac),
-                         Assign(inv_jac, 1/det_jac)]
-            else:
-                body += [Assign(coordinates[i], positions[i][indices_quad[i]]) for i in range(dim)]
-
-            # If assembling a boundary integral, compute any normal/tangent vectors,
-            # as well as the metric determinant for integration on the manifold.
-            if isinstance(self.target, Boundary):
-
-                for vector in vectors:
-
-                    if isinstance(vector, NormalVector):
-                        # replace n[i] by its scalar components
-                        expr  = [e.subs(zip(vector, normal_vec)) for e in expr]
-                        body += compute_normal_vector(locals(), normal_vec, self.target, mapping)
-
-                    elif isinstance(vector, TangentVector):
-                        # replace t[i] by its scalar components
-                        expr  = [e.subs(zip(vector, tangent_vec)) for e in expr]
-                        body += compute_tangent_vector(locals(), tangent_vec, self.target, mapping)
-
-                    else:
-                        raise TypeError(vector)
-
-                body += compute_boundary_jacobian(locals(), self.target, mapping)
-
-            # Evaluation of physical derivatives of basis functions. If there
-            # is no mapping, 'init_map' is an empty list and nothing is done.
-            body += [stmt.subs(1/jac, inv_jac) for stmt in init_map]
-
-            # Evaluation of any scalar and vector fields (and their derivatives)
-            # that appear in the kernel expression.
-            if mapping:
-                body += [Assign(f, v[indices_quad]) for f, v in zip(fields_logical, fields_val)]
-                body += [Assign(f, v[indices_quad])
-                                      for f, v in zip(vector_fields_logical, vector_fields_val)]
-
-                # Convert logical derivatives to physical derivatives
-                body += [stmt.subs(1/jac, inv_jac) for stmt in self._map_stmts_fields.values()]
-
-            else:
-                body += [Assign(f, v[indices_quad]) for f, v in zip(fields, fields_val)]
-                body += [Assign(f, v[indices_quad])
-                                      for f, v in zip(vector_fields, vector_fields_val)]
-
-            # Compute the scalar coefficient V_{ij..} by which the point value
-            # of the kernel expression E must be multiplied in order to
-            # approximate the integral by a high-order quadrature rule:
-            #
-            # integral(domain, E) ~= sum_{ij..} (E_{ij..} * V_{ij..} * J_{ij..}),
-            #
-            # where E_{ij..} = E(x1_i, x2_j, ...), J_{ij..} is the Jacobian
-            # determinant, and V_{ij..} = w1_i * w2_j * ... is simply the
-            # product of the rescaled 1D Gaussian weights along each direction.
-            #
-            # On a boundary with xk=const, we must not multiply by the weight
-            # along xk; hence the need for a 'filter product'.
-            weighted_vol = filter_product(indices_quad, weighted_vols, self.boundary)
-
-            # Multiply by the correct metric determinant
-            # TODO use positive mapping all the time? Abs?
-            if mapping:
-                if isinstance(self.target, Boundary):
-                    weighted_vol *= Abs(det_jac_bnd)
-                else:
-                    weighted_vol *= Abs(det_jac)
-
-            # Update value of integrals:
-            #   1. Assign 'weighted_vol' expression to variable 'wvol',
-            #   2. Multiply each kernel expression by 'wvol',
-            #   3. Increment each integral variables by the expressions above.
-            body.append(Assign(wvol, weighted_vol))
-            for i in range(start, end):
-                if not( i in zero_terms ):
-                    e = SymbolicExpr(Mul(expr[i], wvol))
-                    body.append(AugAssign(vals[i], '+', e))
-
-            #------------------------------------------------------------------
-            # Body of external loop over basis functions
-            #------------------------------------------------------------------
-
-            # Collect initialization statements in this list. These statements
-            # will be placed at the beginning of the function, outside any loops.
-            init_stmts  = []
-
-            # If we are assemblying on a boundary, some indices are set to zero
-            init_stmts += init_loop_quadrature( indices_quad, ranges_quad, self.boundary )
-
-            # Put the body inside for loops of quadrature points
-            body = select_loops( indices_quad, ranges_quad, body, self.boundary,
-                                 boundary_basis=self.boundary_basis)
-
-            # Initialize intermediate variables 'vals': they are set to zero
-            # just before the inner quadrature loop, and they are incremented
-            # inside that loop.
-            init_vars = [Assign(vals[i], 0.0) for i in range(start, end) if not( i in zero_terms )]
-            body = init_vars + body
-
-            if dim_trial:
-                trial_idxs = tuple([indices_trial[i]+trial_pads[i]-indices_test[i] for i in range(dim)])
-                idxs = indices_test + trial_idxs
-            else:
-                idxs = indices_test
-
-            if is_bilinear or is_linear:
-                for i in range(start, end):
-                    if not( i in zero_terms ):
-                        body.append(Assign(mats[i][idxs], vals[i]))
-
-            elif is_function:
-                for i in range(start, end):
-                    if not( i in zero_terms ):
-                        body.append(Assign(mats[i][0], vals[i]))
-
-            #------------------------------------------------------------------
-            # Body of kernel
-            #------------------------------------------------------------------
-
-            # Put the body inside for loops of test and trial functions
-            if is_bilinear:
-                init_stmts += init_loop_basis( indices_test,  ranges_test,  self.boundary )
-                init_stmts += init_loop_basis( indices_trial, ranges_trial, self.boundary )
-
-                body = select_loops(indices_test, ranges_test, body, self.boundary,
-                                    boundary_basis=self.boundary_basis)
-
-                body = select_loops(indices_trial, ranges_trial, body, self.boundary,
-                                    boundary_basis=self.boundary_basis)
-
-            if is_linear:
-                init_stmts += init_loop_basis( indices_test, ranges_test, self.boundary )
-
-                body = select_loops(indices_test, ranges_test, body, self.boundary,
-                                    boundary_basis=self.boundary_basis)
-
+            stmts = []
+            for v in sub_tests+sub_trials:
+                stmts += construct_logical_expressions(v, nderiv)
+            
+            if not mapping.is_analytical:
+                for expr in sub_atomic_expr[:]:
+                    stmts += [ComputePhysicalBasis(i) for i in expr]
             # ...
 
-            # ... add init stmts
-            body = init_stmts + body
+            if atomic_expr_field:
+                stmts += list(eval_fiel.inits)
+
+            loop  = Loop((l_quad, *q_basis_tests.values(), *q_basis_trials.values(), geo), ind_quad, stmts=stmts, mask=mask)
+            loop  = Reduce('+', ComputeKernelExpr(sub_terminal_expr), ElementOf(l_sub_mats), loop)
+
+            # ... loop over trials
+            length = lengths_trials[sub_trials[0]]
+            ind_dof_trial = index_dof_trial.set_length(length)
+            loop  = Loop((*l_basis_trials.values(),), ind_dof_trial, [loop])
+
+            # ... loop over tests
+            length = lengths_tests[sub_tests[0]]
+            ind_dof_test = index_dof_test.set_length(length)
+            loop  = Loop((*l_basis_tests.values(),), ind_dof_test, [loop])
             # ...
 
-            # ...
-            # initialization of the matrix
-            if is_bilinear or is_linear:
-                init_mats = [mats[i][[Slice(None,None)]*(dim_test+dim_trial)] for i in range(start, end) if not( i in zero_terms )]
-
-                init_mats = [Assign(e, 0.0) for e in init_mats]
-                body =  init_mats + body
-
-            # call eval field
-            for eval_field in self.eval_fields:
-                args = test_degrees + basis_test + fields_coeffs + fields_val
-
-                args = eval_field.build_arguments(args)
-
-                body = [FunctionCall(eval_field.func, args)] + body
-
-            # call eval vector_field
-            for eval_vector_field in self.eval_vector_fields:
-                args = test_degrees + basis_test + vector_fields_coeffs + vector_fields_val
-                args = eval_vector_field.build_arguments(args)
-                body = [FunctionCall(eval_vector_field.func, args)] + body
-
-            # call eval mapping
-            if self.eval_mapping:
-                args = (test_degrees + basis_test + mapping_coeffs + mapping_values)
-                args = eval_mapping.build_arguments(args)
-                body = [FunctionCall(eval_mapping.func, args)] + body
-
-            # init/eval area
-            if self.area:
-                # evaluation of the area if the mapping is not used
-                if not mapping:
-                    stmts = [AugAssign(self.area, '+', weighted_vol)]
-                    stmts = select_loops( indices_quad, ranges_quad, stmts, self.boundary,
-                                          boundary_basis=self.boundary_basis)
-
-                    body = stmts + body
-
-                # init area
-                body = [Assign(self.area, 0.0)] + body
-
-            # compute length of logical points
-            len_quads = [Assign(k, Len(u)) for k,u in zip(qds_dim, positions)]
-            body = len_quads + body
-
-            #------------------------------------------------------------------
-            # Compute module-wise import statements
-            #------------------------------------------------------------------
-
-            # Search recursively for math functions and constants
-            def get_math_elements(function_body, lib):
-                math_elements = set()
-                for i in function_body:
-                    if isinstance(i, For):
-                        new = get_math_elements(i.body, lib)
-                        math_elements.update(new)
-                    elif isinstance(i, (Assign, AugAssign)):
-                        new = math_atoms_as_str(i.rhs, lib)
-                        math_elements.update(new)
-                    elif isinstance(i, FunctionCall):
-                        pass
-                    else:
-                        raise TypeError(i)
-                return math_elements
-
-            math_elements = get_math_elements(body, 'numpy')
-            math_imports  = [Import(e, 'numpy') for e in math_elements]
-            self._imports.update(math_imports)
-
-            #------------------------------------------------------------------
-            # Identify function arguments
-            #------------------------------------------------------------------
-
-            mats_args = tuple([mats[i] for i in range(start, end) if not( i in zero_terms )])
-            func_args = fields_coeffs + vector_fields_coeffs + mapping_coeffs + mats_args
-            func_args = self.build_arguments(func_args)
-
-            #------------------------------------------------------------------
-            # Add header and decorators, if needed
-            #------------------------------------------------------------------
-
-            decorators = {}
-            header = None
-            if self.backend['name'] == 'pyccel':
-                decorators = {'types': build_types_decorator(func_args)}
-            elif self.backend['name'] == 'numba':
-                decorators = {'jit':[]}
-            elif self.backend['name'] == 'pythran':
-                header = build_pythran_types_header(self.name, func_args)
-            
-            funcs[i_row][i_col] = FunctionDef(self.name+'_'+str(i_row)+str(i_col), list(func_args), [], body,
-                                    decorators=decorators, header=header)
-
-        return funcs
-
-#==============================================================================
-class Assembly(SplBasic):
-    """
-    Generate the AST of a function for computing an integral form over the
-    whole domain, or the whole boundary. This is obtained by 'assembling' the
-    single contributions over each element (see class Kernel).
-
-    For a bilinear form, such a function will assemble a matrix.
-    For a linear form, it will assemble a vector.
-    For a functional, it will assemble a scalar value.
-
-    """
-    def __new__(cls, kernel, name=None, discrete_space=None, comm=None,
-                mapping=None, is_rational_mapping=None, backend=None):
-
-        # ... Check arguments
-        if not isinstance(kernel, Kernel):
-            raise TypeError('> Expecting a kernel')
-
-        if isinstance(discrete_space, (tuple, list)):
-            space = discrete_space[0]
-        else:
-            space = discrete_space
-
-        if not isinstance(space, (SplineSpace, TensorFemSpace, ProductFemSpace)):
-            raise NotImplementedError('Only Spline, Tensor and Product spaces are available')
-        # ...
-
-        obj = SplBasic.__new__(cls, kernel.tag, name=name,
-                               prefix='assembly', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._kernel         = kernel
-        obj._discrete_space = discrete_space
-        obj._comm           = comm
-        obj._boundary       = kernel.boundary
-        obj._backend        = backend
-
-        # update dependencies
-        obj._dependencies += [kernel]
-
-        obj._func = obj._initialize()
-        return obj
-
-    @property
-    def weak_form(self):
-        return self.kernel.weak_form
-
-    @property
-    def kernel(self):
-        return self._kernel
-
-    @property
-    def discrete_space(self):
-        return self._discrete_space
-
-    @property
-    def comm(self):
-        return self._comm
-
-    @property
-    def global_matrices(self):
-        return self._global_matrices
-
-    @property
-    def backend(self):
-        return self._backend
-
-    def build_arguments(self, data):
-
-        other = data
-
-        if self.kernel.constants:
-            other = other + self.kernel.constants
-
-        if self.kernel.mapping_coeffs:
-            other = self.kernel.mapping_coeffs + other
-
-        return self.basic_args + other
-
-    def _initialize(self):
-        kernel = self.kernel
-        form   = self.weak_form
-        fields = kernel.fields
-        fields_coeffs = kernel.fields_coeffs
-        vector_fields = kernel.vector_fields
-        vector_fields_coeffs = kernel.vector_fields_coeffs
-        zero_terms = kernel.zero_terms
-
-        is_linear   = isinstance(self.weak_form, LinearForm)
-        is_bilinear = isinstance(self.weak_form, BilinearForm)
-        is_function = isinstance(self.weak_form, Functional)
-        
-        if is_bilinear:
-        
-            Wh = self.discrete_space[0]
-            Vh = self.discrete_space[1]
-            is_product_space = isinstance(Wh, ProductFemSpace)
-            ln = 1
-            if is_product_space:
-                ln = len(Wh.spaces)
-        else:
-        
-            Wh = self.discrete_space
-            ln = 1
-            is_product_space = isinstance(self.discrete_space, ProductFemSpace)
-            if is_product_space:
-                ln = len(self.discrete_space.spaces)
-            
-        unique_scalar_space = kernel.unique_scalar_space
-
-        dim    = form.ldim
-
-        n_rows = kernel.n_rows
-        n_cols = kernel.n_cols
-
-        axis_bnd = []
-        if self.boundary:
-            axis_bnd = [self.boundary.axis]
-
-
-        # ... declarations
-
-        starts         = variables('s1:%s'%(dim+1), 'int')
-        ends           = variables('e1:%s'%(dim+1), 'int')
-
-        n_elements     = variables('n_elements_1:%s'%(dim+1), 'int')
-        element_starts = variables('element_s1:%s'%(dim+1),   'int')
-        element_ends   = variables('element_e1:%s'%(dim+1),   'int')
-
-        indices_elm   = variables('ie1:%s'%(dim+1), 'int')
-        indices_span  = variables('is1:%s(1:%s)'%(dim+1, ln+1), 'int')
-
-        test_pads     = variables('test_p1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        trial_pads    = variables('trial_p1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        
-        test_degrees  = variables('test_d1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        trial_degrees = variables('trial_d1:%s(1:%s)'%(dim+1,ln+1), 'int')
-
-        indices_il    = variables('il1:%s'%(dim+1), 'int')
-        indices_i     = variables('i1:%s'%(dim+1),  'int')
-        npts          = variables('n1:%s'%(dim+1),  'int')
-        
-        trial_basis    = variables('trial_basis_1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=4, cls=IndexedVariable)
-        test_basis     = variables('test_basis_1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=4, cls=IndexedVariable)
-
-        spans          = variables('test_spans_1:%s(1:%s)'%(dim+1,ln+1), dtype='int', rank=1, cls=IndexedVariable)
-        quad_orders    = variables( 'k1:%s'%(dim+1), dtype='int')
-
-        trial_basis_in_elm = variables('trial_bs1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=3, cls=IndexedVariable)
-        test_basis_in_elm  = variables('test_bs1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=3, cls=IndexedVariable)
-
-        points_in_elm  = variables('quad_u1:%s'%(dim+1), dtype='real', rank=1, cls=IndexedVariable)
-        weights_in_elm = variables('quad_w1:%s'%(dim+1), dtype='real', rank=1, cls=IndexedVariable)
-
-        points   = variables('points_1:%s'%(dim+1), dtype='real', rank=2, cls=IndexedVariable)
-        weights  = variables('weights_1:%s'%(dim+1), dtype='real', rank=2, cls=IndexedVariable)
-        # ...
-
-        # TODO improve: select args parallel/serial
-        if is_bilinear:
-            self._basic_args = (n_elements +
-                                element_starts + element_ends +
-                                starts + ends +
-                                npts +
-                                quad_orders +
-                                test_degrees + trial_degrees +
-                                test_pads  + trial_pads +
-                                spans +
-                                points + weights +
-                                test_basis + trial_basis)
-
-        if is_linear or is_function:
-            self._basic_args = (n_elements +
-                                element_starts + element_ends +
-                                starts + ends +
-                                npts +
-                                quad_orders +
-                                test_degrees +
-                                test_pads +
-                                spans +
-                                points + weights +
-                                test_basis)
-        # ...
-
-        # ...
-        if is_bilinear:
-            rank = 2*dim
-
-        elif is_linear:
-            rank = dim
-
-        elif is_function:
-            rank = 1
-        # ...
-
-        # ... element matrices
-        element_matrices = OrderedDict()
-        ind = 0
-        for i in range(0, n_rows):
-            for j in range(0, n_cols):
-                if not( ind in zero_terms ):
-                    mat = 'mat_{i}{j}'.format(i=i,j=j)
-
-                    mat = IndexedVariable(mat, dtype='real', rank=rank)
-
-                    element_matrices[i,j] = mat
-
-                ind += 1
-        # ...
-
-        # ... global matrices
-        ind = 0
-        global_matrices = OrderedDict()
-        for i in range(0, n_rows):
-            for j in range(0, n_cols):
-                if not( ind in zero_terms ):
-                    mat = 'M_{i}{j}'.format(i=i,j=j)
-
-                    mat = IndexedVariable(mat, dtype='real', rank=rank)
-
-                    global_matrices[i,j] = mat
-
-                ind += 1
-        # ...
-
-        # sympy does not like ':'
-        _slice = Slice(None,None)
-
-        # assignments
-        body  = [Assign(indices_span[i*ln+j], spans[i*ln+j][indices_elm[i]])
-                 for i,j in np.ndindex(dim, ln) if not(i in axis_bnd)]
-                 
-        if self.debug and self.detailed:
-            msg = lambda x: (String('> span {} = '.format(x)), x)
-            body += [Print(msg(indices_span[i])) for i in range(dim*ln)]
-
-        body += [Assign(points_in_elm[i], points[i][indices_elm[i],_slice])
-                 for i in range(dim) if not(i in axis_bnd) ]
-
-        body += [Assign(weights_in_elm[i], weights[i][indices_elm[i],_slice])
-                 for i in range(dim) if not(i in axis_bnd) ]
-
-        body += [Assign(test_basis_in_elm[i*ln+j], test_basis[i*ln+j][indices_elm[i],_slice,_slice,_slice])
-                 for i,j in np.ndindex(dim,ln) if not(i in axis_bnd) ]
-
-        if is_bilinear:
-            body += [Assign(trial_basis_in_elm[i*ln+j], trial_basis[i*ln+j][indices_elm[i],_slice,_slice,_slice])
-                     for i,j in np.ndindex(dim,ln) if not(i in axis_bnd) ]
-
-        # ... kernel call
-        mats = tuple(element_matrices.values())
-
-        gslices = [Slice(sp-s-d+p, sp+p+1-s) for sp,d,p,s in
-                   zip(indices_span[::ln], test_degrees[::ln], test_pads[::ln], starts)]
-        f_coeffs  = tuple([f[gslices] for f in fields_coeffs])
-        vf_coeffs = tuple([f[gslices] for f in vector_fields_coeffs])
-        m_coeffs  = tuple([f[gslices] for f in kernel.mapping_coeffs])
-
-        if is_bilinear:
-            if not unique_scalar_space:
-                for (i,j), M in element_matrices.items():
-                    args = kernel.build_arguments(f_coeffs + vf_coeffs + m_coeffs + (M,))
-                    args = list(args)
-                    args[:dim] = test_degrees[i::ln]
-                    args[dim:2*dim] = trial_degrees[j::ln]
-                    args[2*dim:3*dim] = [max(pi,pj) for pi,pj in zip(Vh.spaces[i].degree,Wh.spaces[j].degree)]
-                    args[3*dim:4*dim] = test_basis_in_elm[i::ln]
-                    args[4*dim:5*dim] = trial_basis_in_elm[j::ln]
-
-                    body += [FunctionCall(kernel.func[i][j], args)]
-                    
-            else:  
-                args = kernel.build_arguments(f_coeffs + vf_coeffs + m_coeffs + mats)
-                args = list(args)
-
-                args[:dim] = test_degrees[::ln]
-                args[dim:2*dim] = trial_degrees[::ln]
-                args[2*dim:3*dim] = trial_pads[j::ln]
-                args[3*dim:4*dim] = test_basis_in_elm[::ln]
-                args[4*dim:5*dim] = trial_basis_in_elm[::ln]
-
-                body += [FunctionCall(kernel.func[0][0], args)]
-                
-        else:
-            if not unique_scalar_space:
-                for (i,j), M in element_matrices.items():
-                    
-                    args = kernel.build_arguments(f_coeffs + vf_coeffs + m_coeffs + (M,))
-                    args = list(args)
-                    args[:dim] = test_degrees[i::ln]
-                    args[dim:2*dim] = test_basis_in_elm[i::ln]
-                    body += [FunctionCall(kernel.func[i][j], args)]
-                    
-            else:
-                args = kernel.build_arguments(f_coeffs + vf_coeffs + m_coeffs + mats)
-                args = list(args)
-                args[:dim] = test_degrees[::ln]
-                args[dim:2*dim] = test_basis_in_elm[::ln]
-                body += [FunctionCall(kernel.func[0][0], args)]
-            
-        # ...
-
-        # ... update global matrices
-        lslices = [Slice(None,None)]*dim
-        if is_bilinear:
-            lslices += [Slice(None,None)]*dim # for assignement
-
-
-        if is_function:
-            lslices = 0
-            gslices = 0
-
-        for (i,j), M in global_matrices.items():
-           
-            mat = element_matrices[i,j]
-            local_test_degrees = test_degrees[i::ln]
-            local_indices_span = indices_span[i::ln]
-            local_test_pads = test_pads[i::ln]
-                
-
-            if is_bilinear:
-
-                if ( self.comm is None ):
-                    gslices = [Slice(sp-d+p,sp+p+1) for sp,d,p in zip(local_indices_span, local_test_degrees,local_test_pads)]
+            body  = (Reset(l_sub_mats), loop)
+            stmts = Block(body)
+            g_stmts += [stmts]
     
-                else:
-                    gslices = [Slice(sp-s-d+p,sp+p+1-s) for sp,d,p,s in zip(local_indices_span,
-                                                                   local_test_degrees,
-                                                                   local_test_pads,
-                                                                       starts)]
+    #=========================================================end kernel=========================================================
 
-                gslices += [Slice(None,None)]*dim # for assignement
+    # ... loop over global elements
+    loop  = Loop((g_quad, *g_basis_tests.values(), *g_basis_trials.values(), *g_span.values()),
+                  ind_element, stmts=g_stmts, mask=mask)
 
-            if is_linear:
-                if ( self.comm is None ):
-                    gslices = [Slice(sp-d+p,sp+p+1) for sp,d,p in zip(local_indices_span, local_test_degrees,local_test_pads)]
+    body = [Reduce('+', l_mats, g_mats, loop)]
+    # ...
+    args = OrderedDict()
+    args['tests_basis']  = g_basis_tests.values()
+    args['trial_basis']  = g_basis_trials.values()
 
-                else:
-                    gslices = [Slice(sp-s-d+p,sp+p+1-s) for sp,d,p,s in zip(local_indices_span,
-                                                                   local_test_degrees,
-                                                                   local_test_pads,
-                                                                   starts)]
-            
-            stmt = AugAssign(M[gslices], '+', mat[lslices])
+    args['spans'] = g_span.values()
+    args['quads'] = g_quad
 
-            body += [stmt]
+    args['tests_degrees'] = lengths_tests
+    args['trials_degrees'] = lengths_trials
+
+    args['quads_degree'] = lengths
+    args['global_pads']  = pads
+    args['local_pads']   = Pads(tests, trials)
+
+    args['mats']  = [l_mats, g_mats]
+
+    if eval_mapping:
+        args['mapping'] = eval_mapping.coeffs
+
+    if atomic_expr_field:
+        args['coeffs'] = l_coeffs
+
+    local_vars = [*q_basis_tests, *q_basis_trials]
+    stmt = DefNode('assembly', args, local_vars, body)
+
+    return stmt
+
+#================================================================================================================================
+def _create_ast_linear_form(terminal_expr, atomic_expr, atomic_expr_field, tests, d_tests, shapes_tests, nderiv, dim, mapping, space, mask, tag):
+    """
+    """
+    pads   = symbols('pad1, pad2, pad3')[:dim]
+    g_quad = GlobalTensorQuadrature()
+    l_quad = LocalTensorQuadrature()
+
+    coeffs   = [CoefficientBasis(i) for i in tests]
+    l_coeffs = [MatrixLocalBasis(i) for i in tests]
+    geo      = GeometryExpressions(mapping, nderiv)
+
+    l_vecs  = BlockStencilVectorLocalBasis(tests, pads, terminal_expr, tag)
+    g_vecs  = BlockStencilVectorGlobalBasis(tests, pads, terminal_expr,l_vecs.tag)
+
+    q_basis  = OrderedDict((v,d_tests[v]['array'])  for v in tests)
+    l_basis  = OrderedDict((v,d_tests[v]['local'])  for v in tests)
+    g_basis  = OrderedDict((v,d_tests[v]['global'])  for v in tests)
+
+    g_span          = OrderedDict((v,d_tests[v]['span']) for v in tests)
+    lengths_tests   = OrderedDict((v,LengthDofTest(v)) for v in tests)
+    # ...........................................................................................
+    quad_length = LengthQuadrature()
+    el_length   = LengthElement()
+    lengths     = [el_length,quad_length]
+
+    ind_quad      = index_quad.set_length(quad_length)
+    ind_element   = index_element.set_length(el_length)
+    ind_dof_test = index_dof_test.set_length(LengthDofTest(tests[0]))
+    # ...........................................................................................
+    eval_mapping = EvalMapping(ind_quad, ind_dof_test, q_basis[tests[0]], l_basis[tests[0]], mapping, geo, space, tests, nderiv, mask)
+
+    if atomic_expr_field:
+        raise NotImplementedError('TODO')
+        eval_field   = EvalField(atomic_expr_field, ind_quad, q_basis, coeff, tests, mapping, nderiv)
+
+    # ...
+    g_stmts = [eval_mapping]
+    groups = regroupe(tests)
+    ex_tests = expand(tests)
+    # ... 
+    #=========================================================begin kernel======================================================
+
+    for dtype, group in groups:
+        tests_indices = [ex_tests.index(i) for i in expand(group)]
+        sub_terminal_expr = terminal_expr[tests_indices,0]
+        sub_atomic_expr   = atomic_expr[tests_indices,0]
+        l_sub_vecs  = BlockStencilVectorLocalBasis(group, pads, sub_terminal_expr, l_vecs.tag)
+        q_basis = {v:d_tests[v]['array']  for v in group}
+        l_basis = {v:d_tests[v]['local']  for v in group}
+        if sub_terminal_expr.is_zero:
+            continue
+        stmts = []
+        for v in group:
+            stmts += construct_logical_expressions(v, nderiv)
+
+        if mapping._expressions is None:
+            for expr in sub_atomic_expr[:]:
+                stmts += [ComputePhysicalBasis(i) for i in expr]
+
+        if atomic_expr_field:
+            stmts += list(eval_fiel.inits)
+
+        loop  = Loop((l_quad, *q_basis.values(), geo), ind_quad, stmts=stmts, mask=mask)
+        loop = Reduce('+', ComputeKernelExpr(sub_terminal_expr), ElementOf(l_sub_vecs), loop)
+    # ...
+
+    # ... loop over tests to evaluate fields
+        fields = EmptyLine()
+        if atomic_expr_field:
+            fields = Loop((*l_basis, l_coeff), ind_dof, [eval_field])
+
+    # ... loop over tests
+        length   = lengths_tests[group[0]]
+        ind_dof_test = index_dof_test.set_length(length)
+        loop  = Loop((*l_basis.values(),), ind_dof_test, [loop])
         # ...
+        body  = (EmptyLine(), fields, Reset(l_sub_vecs), loop)
+        stmts = Block(body)
+        g_stmts += [stmts]
+    # ...
+    
+    #=========================================================end kernel=========================================================
+    # ... loop over global elements
+    loop  = Loop((g_quad, *g_basis.values(), *g_span.values()), ind_element, stmts=g_stmts, mask=mask)
+    # ...
+    body = (Reduce('+', l_vecs, g_vecs, loop),)
 
-        # ... loop over elements
-        if is_function:
-            ranges_elm  = [Range(s, e+1) for s,e in zip(element_starts,
-                                                      element_ends)]
-
-        else:
-            ranges_elm  = [Range(0, n_elements[i]) for i in range(dim)]
-
-        # TODO call init_loops
-        init_stmts = init_loop_support( indices_elm, n_elements,
-                                       indices_span, spans, ranges_elm,
-                                       points_in_elm, points,
-                                       weights_in_elm, weights,
-                                       test_basis_in_elm, test_basis,
-                                       trial_basis_in_elm, trial_basis,
-                                       is_bilinear, self.boundary )
-
-        body = select_loops(indices_elm, ranges_elm, body,
-                            self.kernel.boundary, boundary_basis=False)
+    args = OrderedDict()
+    args['tests_basis']  = g_basis.values()
 
-        body = init_stmts + body
-
-        # ...
+    args['spans'] = g_span.values()
+    args['quads'] = g_quad
 
-        # ... prelude
-        imports = []
-
-        # import zeros from numpy
-        stmt = Import('zeros', 'numpy')
-        imports += [stmt]
+    args['tests_degrees'] = lengths_tests
 
-        # import product from itertools
-        stmt = Import('product', 'itertools')
-        imports += [stmt]
+    args['quads_degree'] = lengths
+    args['global_pads']  = pads
 
-        prelude = []
-        # allocate element matrices
+    args['mats']  = [l_vecs, g_vecs]
 
-        for (i,j),mat in element_matrices.items():
+    if eval_mapping:
+        args['mapping'] = eval_mapping.coeffs
 
-            orders  = [p+1 for p in test_degrees[i::ln]]
-            spads   = [2*p+1 for p in test_pads[j::ln]]
-            
-            if is_bilinear:
-                if not unique_scalar_space:
-                    spads = [2*max(pi,pj)+1 for pi,pj in zip(Vh.spaces[i].degree,Wh.spaces[j].degree)]
-                    
-                args  = tuple(orders + spads)
+    if atomic_expr_field:
+        args['coeffs'] = l_coeffs
 
-            if is_linear:
-                args = tuple(orders)
+    local_vars = [*q_basis]
+    stmt = DefNode('assembly', args, local_vars, body)
+    # ...
 
-            if is_function:
-                args = tuple([1])
+    return stmt
 
-            stmt = Assign(mat, Zeros(args))
-            prelude += [stmt]
+#================================================================================================================================
+def _create_ast_functional_form(terminal_expr, atomic_expr, tests, d_tests, shapes_tests, nderiv, dim, mapping, space, mask, tag):
+    """
+    """
 
+    pads   = symbols('pad1, pad2, pad3')[:dim]
+    g_quad = GlobalTensorQuadrature()
+    l_quad = LocalTensorQuadrature()
 
+    #TODO move to EvalField
+    coeffs   = [CoefficientBasis(i) for i in expand(tests)]
+    l_coeffs = [MatrixLocalBasis(i) for i in expand(tests)]
+    g_coeffs = [MatrixGlobalBasis(i,i) for i in expand(tests)]
 
-        # allocate mapping values
-        if self.kernel.mapping_values:
-            for v in self.kernel.mapping_values:
-                stmt = Assign(v, Zeros(quad_orders))
-                prelude += [stmt]
+    geo      = GeometryExpressions(mapping, nderiv)
 
-        # TODO allocate field values
-        if self.kernel.fields:
-            fields_shape = tuple(FunctionCall('len',[p[0,Slice(None,None)]]) for p in points)
-            for F_value in self.kernel.fields_val:
-                prelude += [Assign(F_value, Zeros(fields_shape))]
+    q_basis  = OrderedDict((v,d_tests[v]['array'])  for v in tests)
+    l_basis  = OrderedDict((v,d_tests[v]['local'])  for v in tests)
+    g_basis  = OrderedDict((v,d_tests[v]['global'])  for v in tests)
+    g_span          = OrderedDict((v,d_tests[v]['span']) for v in tests)
+    lengths_tests   = OrderedDict((v,LengthDofTest(v)) for v in tests)
 
-        if self.kernel.vector_fields_val:
-            fields_shape = tuple(FunctionCall('len',[p[0,Slice(None,None)]]) for p in points)
-            for F_value in self.kernel.vector_fields_val:
-                prelude += [Assign(F_value, Zeros(fields_shape))]
+    l_vec   = LocalElementBasis()
+    g_vec   = GlobalElementBasis()
 
-        # ...
-        if self.debug:
-            for ij, M in global_matrices.items():
-                i,j = ij
-                prelude += [Print((String('> shape {} = '.format(M)), Shape(M)))]
-        # ...
+    # ...........................................................................................
+    quad_length = LengthQuadrature()
+    el_length   = LengthElement()
+    lengths     = [el_length, quad_length]
 
-        # ...
-        body = prelude + body
-        # ...
+    ind_quad      = index_quad.set_length(quad_length)
+    ind_element   = index_element.set_length(el_length)
+    ind_dof_test  = index_dof_test.set_length(lengths_tests[tests[0]])
+    # ...........................................................................................
+    eval_mapping = EvalMapping(ind_quad, ind_dof_test, q_basis[tests[0]], l_basis[tests[0]], mapping, geo, space, tests, nderiv, mask)
+    eval_field   = EvalField(atomic_expr, ind_quad, ind_dof_test, q_basis[tests[0]], l_basis[tests[0]], coeffs, l_coeffs, g_coeffs, tests, mapping, pads, nderiv, mask)
 
-        # ...
-        mats = tuple(global_matrices.values())
-        self._global_matrices = global_matrices
-        # ...
-        self._imports = imports
-        # function args
-        func_args = self.build_arguments(fields_coeffs + vector_fields_coeffs + mats)
+    #=========================================================begin kernel======================================================
+    # ... loop over tests functions
 
-        decorators = {}
-        header = None
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args),'external_call':[]}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
+    loop   = Loop((l_quad, geo), ind_quad, eval_field.inits)
+    loop   = Reduce('+', ComputeKernelExpr(terminal_expr), ElementOf(l_vec), loop)
 
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
+    # ... loop over tests functions to evaluate the fields
 
+    stmts  = Block([eval_mapping, eval_field, Reset(l_vec), loop])
 
-#==============================================================================
-class Interface(SplBasic):
+    #=========================================================end kernel=========================================================
+    # ... loop over global elements
 
-    def __new__(cls, assembly, name=None, backend=None,
-                discrete_space=None, comm=None, mapping=None, is_rational_mapping=None):
 
-        if not isinstance(assembly, Assembly):
-            raise TypeError('> Expecting an Assembly')
-
-        obj = SplBasic.__new__(cls, assembly.tag, name=name,
-                               prefix='interface', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._assembly = assembly
-        obj._backend = backend
-        obj._discrete_space = discrete_space
-        obj._comm = comm
-
-        dim = assembly.weak_form.ldim
-
-
-        # update dependencies
-	# TODO uncomment later
-        #lo_dot = LinearOperatorDot(dim, backend)
-        #v_dot  = VectorDot(dim, backend)
-
-        #obj._dots = [lo_dot, v_dot]
-        #obj._dependencies += [assembly, lo_dot, v_dot]
-
-        obj._dependencies += [assembly]
-
-        obj._func = obj._initialize()
-        return obj
-
-    @property
-    def weak_form(self):
-        return self._assembly.weak_form
-        
-    @property
-    def space(self):
-        return self._assembly.kernel.symbolic_space
-
-    @property
-    def assembly(self):
-        return self._assembly
-
-    @property
-    def backend(self):
-        return self._backend
-
-    @property
-    def discrete_space(self):
-        return self._discrete_space
-
-    @property
-    def comm(self):
-        return self._comm
-
-    @property
-    def max_nderiv(self):
-        return self.assembly.kernel.max_nderiv
-
-    def build_arguments(self, data):
-        # data must be at the end, since they are optional
-        return self.basic_args + data
-
-    @property
-    def in_arguments(self):
-        return self._in_arguments
-
-    @property
-    def inout_arguments(self):
-        return self._inout_arguments
-
-    @property
-    def user_functions(self):
-        return self.assembly.kernel.user_functions
-
-# TODO uncomment later
-    #@property
-    #def dots(self):
-    #    return self._dots
-
-
-    def _initialize(self):
-        form = self.weak_form
-        assembly = self.assembly
-        global_matrices = assembly.global_matrices
-        fields = tuple(form.expr.atoms(ScalarField))
-        fields = sorted(fields, key=lambda x: str(x.name))
-        fields = tuple(fields)
-        zero_terms = assembly.kernel.zero_terms
-
-        vector_fields = tuple(form.expr.atoms(VectorField))
-        vector_fields = sorted(vector_fields, key=lambda x: str(x.name))
-        vector_fields = tuple(vector_fields)
-
-        is_linear   = isinstance(self.weak_form, LinearForm)
-        is_bilinear = isinstance(self.weak_form, BilinearForm)
-        is_function = isinstance(self.weak_form, Functional)
-
-        dim = form.ldim
-        
-        if is_bilinear:
-            Wh = self.discrete_space[0]
-            Vh = self.discrete_space[1]
-            
-        else:
-            Wh = self.discrete_space
-            
-        is_product_fem_space = isinstance(Wh, ProductFemSpace)
-        unique_scalar_space = assembly.kernel.unique_scalar_space
-
-        # ... declarations
-
-        test_space = Symbol('W')
-        trial_space = Symbol('V')
-        grid = Symbol('grid')
-        test_basis_values = Symbol('test_basis_values')
-        trial_basis_values = Symbol('trial_basis_values')
-
-        if is_bilinear:
-            basis_values = (test_basis_values, trial_basis_values)
-
-        else:
-            basis_values = (test_basis_values,)
-
-        if is_bilinear:
-            spaces = (test_space, trial_space)
-            test_vector_space = DottedName(test_space, 'vector_space')
-            trial_vector_space = DottedName(trial_space, 'vector_space')
-
-            ln = 1
-            if is_product_fem_space: 
-                ln = len(Wh.spaces)         
-                test_vector_space = DottedName(test_vector_space, 'spaces')
-                trial_vector_space = DottedName(trial_vector_space, 'spaces')
-            # ...
-
-        if is_linear or is_function:
-            test_vector_space = DottedName(test_space, 'vector_space')
-            spaces = (test_space,)
-            ln = 1
-            if is_product_fem_space:
-                ln = len(Wh.spaces)
-                test_vector_space = DottedName(test_vector_space, 'spaces')
-            # ...
-
-        n_elements     = variables('n_elements_1:%s'%(dim+1), 'int')
-        starts         = variables('s1:%s'%(dim+1), 'int')
-        ends           = variables('e1:%s'%(dim+1), 'int')
-        npts           = variables('n1:%s'%(dim+1), 'int')
-        element_starts = variables('element_s1:%s'%(dim+1), 'int')
-        element_ends   = variables('element_e1:%s'%(dim+1), 'int')
-
-        test_degrees   = variables('test_d1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        trial_degrees  = variables('trial_d1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        
-        test_pads      = variables('test_p1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        trial_pads     = variables('trial_p1:%s(1:%s)'%(dim+1,ln+1), 'int')
-        
-        trial_basis    = variables('trial_basis_1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=4, cls=IndexedVariable)
-        test_basis     = variables('test_basis_1:%s(1:%s)'%(dim+1,ln+1), dtype='real', rank=4, cls=IndexedVariable)
-
-        spans          = variables('test_spans_1:%s(1:%s)'%(dim+1,ln+1), dtype='int', rank=1, cls=IndexedVariable)
-        quad_orders    = variables( 'k1:%s'%(dim+1), dtype='int')
-
-        points         = variables('points_1:%s'%(dim+1),  dtype='real', rank=2, cls=IndexedVariable)
-        weights        = variables('weights_1:%s'%(dim+1), dtype='real', rank=2, cls=IndexedVariable)
-
-
-        test_spaces, trial_spaces = symbols('test_spaces, trial_spaces', cls=IndexedBase)
-        spans_attr , basis_attr   = symbols('spans, basis', cls=IndexedBase)
-        pads                      = symbols('pads')
-        
-	# TODO uncomment later
-        #dots           = symbols('lo_dot v_dot')
-        #dot            = Symbol('dot')
-
-        mapping = ()
-        if self.mapping:
-            mapping = Symbol('mapping')
-        # ...
-
-        # ...
-        self._basic_args = spaces + (grid,) + basis_values
-        # ...
-
-        spaces = IndexedBase('spaces')
-        
-        # ... interface body
-        body = []
-        body += [Assign(test_spaces, test_vector_space)]
-        
-        if is_bilinear:
-            body += [Assign(trial_spaces, trial_vector_space)]
-            
-        # ... grid data
-        body += [Assign(n_elements,     DottedName(grid, 'n_elements'))]
-        body += [Assign(points,         DottedName(grid, 'points'))]
-        body += [Assign(weights,        DottedName(grid, 'weights'))]
-        body += [Assign(quad_orders,    DottedName(grid, 'quad_order'))]
-        body += [Assign(element_starts, DottedName(grid, 'local_element_start'))]
-        body += [Assign(element_ends,   DottedName(grid, 'local_element_end'))]
-        # ...
-
-        # ... basis values
-        if is_product_fem_space:
-            for i in range(ln):
-                body += [Assign(spans[i::ln],      DottedName(test_basis_values, spans_attr[i]))]
-                body += [Assign(test_basis[i::ln], DottedName(test_basis_values, basis_attr[i]))]   
-       
-        else:
-            body += [Assign(spans,      DottedName(test_basis_values, spans_attr))]
-            body += [Assign(test_basis, DottedName(test_basis_values, basis_attr))]
-
-        if is_bilinear:
-            if is_product_fem_space:
-                for i in range(ln):
-                    body += [Assign(trial_basis[i::ln], DottedName(trial_basis_values, basis_attr[i]))]   
-            else:
-                body += [Assign(trial_basis, DottedName(trial_basis_values, basis_attr))]
-        # ...
-
-        # ... getting data from fem space
-        if is_product_fem_space:
-            for i in range(ln):
-                body += [Assign(test_degrees[i::ln], DottedName(test_space,spaces[i], 'degree'))]
-                body += [Assign(test_pads   [i::ln], DottedName(test_spaces[i], 'pads'))]
-            
-        else:
-            body += [Assign(test_degrees, DottedName(test_space, 'degree'))]
-            body += [Assign(test_pads   , DottedName(test_spaces, 'pads'))]
-            
-        if is_bilinear:
-        
-            if is_product_fem_space:
-                for i in range(ln):
-                    body += [Assign(trial_degrees[i::ln], DottedName(trial_space,spaces[i], 'degree'))]
-                    body += [Assign(trial_pads   [i::ln], DottedName(trial_spaces[i], 'pads'))]
-            else:
-                body += [Assign(trial_degrees, DottedName(trial_space, 'degree'))]
-                body += [Assign(trial_pads   , DottedName(trial_spaces, 'pads'))]
-
-        if is_product_fem_space:
-            body += [Assign(starts, DottedName(test_spaces[0], 'starts'))]
-            body += [Assign(ends,   DottedName(test_spaces[0], 'ends'))]
-            body += [Assign(npts,   DottedName(test_spaces[0], 'npts'))]
-        
-        else:
-            body += [Assign(starts, DottedName(test_spaces, 'starts'))]
-            body += [Assign(ends,   DottedName(test_spaces, 'ends'))]
-            body += [Assign(npts,   DottedName(test_spaces, 'npts'))]
-        # ...
-        if mapping:
-            # we limit the range to dim, since the last element can be the
-            # weights when using NURBS
-            for i, coeff in enumerate(assembly.kernel.mapping_coeffs[:dim]):
-                component = IndexedBase(DottedName(mapping, '_fields'))[i]
-                c_var = DottedName(component, '_coeffs', '_data')
-                body += [Assign(coeff, c_var)]
-
-            # NURBS case
-            if self.is_rational_mapping:
-                coeff = assembly.kernel.mapping_coeffs[-1]
-
-                component = DottedName(mapping, '_weights_field')
-                c_var = DottedName(component, '_coeffs', '_data')
-                body += [Assign(coeff, c_var)]
-        # ...
-
-        # ...
-        imports = []
-        if not is_function:
-            if is_bilinear:
-                imports += [Import('StencilMatrix', 'psydac.linalg.stencil')]
-
-            if is_linear:
-                imports += [Import('StencilVector', 'psydac.linalg.stencil')]
-
-
-            for ij,M in global_matrices.items():
-                (i,j) = ij
-                if_cond = Is(M, Nil())
-                if is_bilinear:
-                    if is_product_fem_space:
-                        spj = Vh.spaces[j]
-                        spi = Wh.spaces[i]
-                        pads_args = tuple(max(pi,pj) for pi,pj in zip(spj.degree,spi.degree))
-                        args = [trial_spaces[j], test_spaces[i], Assign(pads,pads_args)]
-                    else:
-                        args = [trial_spaces , test_spaces]
-                        
-                    if_body = [Assign(M, FunctionCall('StencilMatrix', args))]
-                    # TODO uncomment later
-                    #if_body.append(Assign(DottedName(M,'_dot'),dots[0]))
-
-                if is_linear:
-                    if is_product_fem_space:
-                        args = [test_spaces[i]]
-                    else:
-                        args = [test_spaces]
-                        
-                    if_body = [Assign(M, FunctionCall('StencilVector', args))]
-                    # TODO uncomment later
-                    #if_body.append(Assign(DottedName(M,'_dot'),dots[1]))
-
-                stmt = If((if_cond, if_body))
-                body += [stmt]
-
-        else:
-            imports += [Import('zeros', 'numpy')]
-            for M in global_matrices.values():
-                body += [Assign(M, Zeros(1))]
-        # ...
-
-        # ...
-        self._inout_arguments = list(global_matrices.values())
-        self._in_arguments = list(self.assembly.kernel.constants) + list(fields) + list(vector_fields)
-        # ...
-
-        # ... call to assembly
-        if is_bilinear or is_linear:
-            mat_data = [DottedName(M, '_data') for M in global_matrices.values()]
-
-        elif is_function:
-            mat_data = [M for M in global_matrices.values()]
-
-        mat_data       = tuple(mat_data)
-
-        field_data     = [DottedName(F, '_coeffs', '_data') for F in fields]
-        field_data     = tuple(field_data)
-
-        vector_field_data     = [DottedName(F, '_coeffs[{}]'.format(i),
-                                            '_data') for F in
-                                 vector_fields for i in range(0, dim)]
-        vector_field_data     = tuple(vector_field_data)
-
-        args = assembly.build_arguments(field_data + vector_field_data + mat_data)
-
-        body += [FunctionCall(assembly.func, args)]
-        # ...
-
-        # ... IMPORTANT: ghost regions must be up-to-date
-        if not( self.comm is None ):
-            if is_linear:
-                for M in global_matrices.values():
-                    f_name = '{}.update_ghost_regions'.format(str(M.name))
-                    stmt = FunctionCall(f_name, [])
-                    body += [stmt]
-        # ...
-
-        # ... results
-        if is_bilinear or is_linear:
-            n_rows = self.assembly.kernel.n_rows
-            n_cols = self.assembly.kernel.n_cols
-
-            if n_rows * n_cols > 1:
-                if is_bilinear:
-                    L = IndexedBase('L')
-
-                    imports += [Import('BlockMatrix', 'psydac.linalg.block')]
-
-                    # ... TODO this is a duplicated code => use a function to define
-                    # global_matrices
-                    ind = 0
-                    d = {}
-                    for i in range(0, n_rows):
-                        for j in range(0, n_cols):
-                            if not( ind in zero_terms ):
-                                mat = IndexedBase('M_{i}{j}'.format(i=i,j=j))
-                                d[(i,j)] = mat
-
-                            ind += 1
-                    # ...
-
-                    # ... create product space
-                    test_vector_space  = DottedName(test_space , 'vector_space')
-                    trial_vector_space = DottedName(trial_space, 'vector_space')
-                    body += [Assign(L, FunctionCall('BlockMatrix', [test_vector_space, trial_vector_space]))]
-                    d = OrderedDict(sorted(d.items()))
-                    for k,v in d.items():
-                        body += [Assign(L[k], v)]
-
-
-                elif is_linear:
-                    L = IndexedBase('L')
-
-                    # ... TODO this is a duplicated code => use a function to define
-                    # global_matrices
-                    # n_cols is equal to 1
-
-                    ind = 0
-                    d = {}
-                    j = 0
-                    for i in range(0, n_rows):
-                        if not( ind in zero_terms ):
-                            mat = IndexedBase('M_{i}{j}'.format(i=i,j=j))
-                            d[i] = mat
-
-                        ind += 1
-                    # ...
-
-                    imports += [Import('BlockVector', 'psydac.linalg.block')]
-
-                    # ... create product space
-                    test_vector_space = DottedName(test_space, 'vector_space')
-                    # ...
-
-                    body += [Assign(L, FunctionCall('BlockVector', [test_vector_space]))]
-                    d = OrderedDict(sorted(d.items()))
-                    for k,v in d.items():
-                        body += [Assign(L[k], v)]
-
-                body += [Return(L)]
-
-            else:
-                M = list(global_matrices.values())[0]
-                body += [Return(M)]
-
-        elif is_function:
-            if len(global_matrices) == 1:
-                M = list(global_matrices.values())[0]
-                body += [Return(M[0])]
-
-            else:
-                body += [Return(M[0]) for M in global_matrices.values()]
-        # ...
-
-        # ... arguments
-        if is_bilinear or is_linear:
-            mats = [Assign(M, Nil()) for M in global_matrices.values()]
-            mats = tuple(mats)
-
-        elif is_function:
-            mats = ()
-
-        if mapping:
-            mapping = (mapping,)
-
-        if self.assembly.kernel.constants:
-            constants = self.assembly.kernel.constants
-            args = mapping + constants + fields + vector_fields + mats
-
-        else:
-            args = mapping + fields + vector_fields + mats
-
-        func_args = self.build_arguments(args)
-        # ...
-
-        self._imports = imports
-        
-        return FunctionDef(self.name, list(func_args), [], body)
-
-
-# TODO uncomment later
-class LinearOperatorDot(SplBasic):
-
-    def __new__(cls, ndim, backend=None):
-
-
-        obj = SplBasic.__new__(cls, 'dot',name='lo_dot',prefix='lo_dot')
-        obj._ndim = ndim
-        obj._backend = backend
-        obj._func = obj._initilize()
-        return obj
-
-    @property
-    def ndim(self):
-        return self._ndim
-
-    @property
-    def func(self):
-        return self._func
-
-    @property
-    def backend(self):
-        return self._backend
-
-
-    def _initilize(self):
-
-        ndim = self.ndim
-        nrows           = variables('n1:%s'%(ndim+1),  'int')
-        pads            = variables('p1:%s'%(ndim+1),  'int')
-        indices1        = variables('ind1:%s'%(ndim+1),'int')
-        indices2        = variables('i1:%s'%(ndim+1),  'int')
-        extra_rows      = variables('extra_rows','int',rank=1,cls=IndexedVariable)
-
-        ex,v            = variables('ex','int'), variables('v','real')
-        x, out          = variables('x, out','real',cls=IndexedVariable, rank=ndim)
-        mat             = variables('mat','real',cls=IndexedVariable, rank=2*ndim)
-
-        body = []
-        ranges = [Range(2*p+1) for p in pads]
-        target = Product(*ranges)
-
-
-        v1 = x[tuple(i+j for i,j in zip(indices1,indices2))]
-        v2 = mat[tuple(i+j for i,j in zip(indices1,pads))+tuple(indices2)]
-        v3 = out[tuple(i+j for i,j in zip(indices1,pads))]
-
-        body = [AugAssign(v,'+' ,Mul(v1,v2))]
-        body = [For(indices2, target, body)]
-        body.insert(0,Assign(v, 0.0))
-        body.append(Assign(v3,v))
-        ranges = [Range(i) for i in nrows]
-        target = Product(*ranges)
-        body = [For(indices1,target,body)]
-
-        for dim in range(ndim):
-            body.append(Assign(ex,extra_rows[dim]))
-
-
-            v1 = [i+j for i,j in zip(indices1, indices2)]
-            v2 = [i+j for i,j in zip(indices1, pads)]
-            v1[dim] += nrows[dim]
-            v2[dim] += nrows[dim]
-            v3 = v2
-            v1 = x[tuple(v1)]
-            v2 = mat[tuple(v2)+ indices2]
-            v3 = out[tuple(v3)]
-
-            rows = list(nrows)
-            rows[dim] = ex
-            ranges = [2*p+1 for p in pads]
-            ranges[dim] -= indices1[dim] + 1
-            ranges =[Range(i) for i in ranges]
-            target = Product(*ranges)
-
-            for_body = [AugAssign(v, '+',Mul(v1,v2))]
-            for_body = [For(indices2, target, for_body)]
-            for_body.insert(0,Assign(v, 0.0))
-            for_body.append(Assign(v3,v))
-
-            ranges = [Range(i) for i in rows]
-            target = Product(*ranges)
-            body += [For(indices1, target, for_body)]
-
-
-        func_args =  (extra_rows, mat, x, out) + nrows + pads
-
-        self._imports = [Import('product','itertools')]
-
-        decorators = {}
-        header = None
-
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args), 'external_call':[]}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
-
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
-
-
-class VectorDot(SplBasic):
-
-    def __new__(cls, ndim, backend=None):
-
-
-        obj = SplBasic.__new__(cls, 'dot', name='v_dot', prefix='v_dot')
-        obj._ndim = ndim
-        obj._backend = backend
-        obj._func = obj._initilize()
-        return obj
-
-    @property
-    def ndim(self):
-        return self._ndim
-
-    @property
-    def func(self):
-        return self._func
-
-    @property
-    def backend(self):
-        return self._backend
-
-    def _initilize(self):
-
-        ndim = self.ndim
-
-        indices = variables('i1:%s'%(ndim+1),'int')
-        dims    = variables('n1:%s'%(ndim+1),'int')
-        pads    = variables('p1:%s'%(ndim+1),'int')
-        out     = variables('out','real')
-        x1,x2   = variables('x1, x2','real',rank=ndim,cls=IndexedVariable)
-
-        body = []
-        ranges = [Range(p,n-p) for n,p in zip(dims,pads)]
-        target = Product(*ranges)
-
-
-        v1 = x1[indices]
-        v2 = x2[indices]
-
-        body = [AugAssign(out,'+' ,Mul(v1,v2))]
-        body = [For(indices, target, body)]
-        body.insert(0,Assign(out, 0.0))
-        body.append(Return(out))
-
-        func_args =  (x1, x2) + pads + dims
-
-        self._imports = [Import('product','itertools')]
-
-        decorators = {}
-        header = None
-
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args), 'external':[]}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
-
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
+    loop  = Loop((g_quad, *g_basis.values(), *g_span.values()), ind_element, stmts)
+    # ...
+
+    body = (Reduce('+', l_vec, g_vec, loop),)
+
+    args = OrderedDict()
+    args['tests_basis']  = g_basis.values()
+
+    args['spans'] = g_span.values()
+    args['quads'] = g_quad
+
+    args['tests_degrees'] = lengths_tests
+
+    args['quads_degree'] = lengths
+    args['global_pads']  = pads
+
+    args['mats']  = [l_vec, g_vec]
+
+    if eval_mapping:
+        args['mapping'] = eval_mapping.coeffs
+
+    args['coeffs'] = g_coeffs
+
+    local_vars = [*q_basis]
+    stmt = DefNode('assembly', args, local_vars, body)
+    # ...
+
+    return stmt
