--- psydac/api/new_ast/__init__.py
@@ -1,5 +1,0 @@
-from psydac.api.new_ast import nodes
-from psydac.api.new_ast import parser
-from psydac.api.new_ast import utilities
-from psydac.api.new_ast import fem
-
