+++ psydac/api/old_ast/__init__.py
@@ -1,0 +1,4 @@
+from psydac.api.old_ast import fem
+from psydac.api.old_ast import basic
+from psydac.api.old_ast import evaluation
+from psydac.api.old_ast import utilities
